// src/app/shared/types/api-types/bidt-package-dto.interface.ts
// //////////////////////////////////////////////////////////////////////////
// DO NOT EDIT THIS FILE
//
// It is automatically generated by generate-dtos.js
// //////////////////////////////////////////////////////////////////////////

import { BidtDfItemDTO } from './bidt-df-item-dto.interface';
import { BidtDocumentDTO } from './bidt-document-dto.interface';
import { BidtPackageIdDTO } from './bidt-package-id-dto.interface';
import { BidtPackageOperationDTO } from './bidt-package-operation-dto.interface';

export interface BidtPackageDTO extends BidtPackageIdDTO {
  id?: number;
  bidtDfItems?: BidtDfItemDTO[];
  bidtDocuments?: BidtDocumentDTO[];
  bidtPackageOperations?: BidtPackageOperationDTO[];
  packageCode?: string; // 50
  packageStatus?: string; // 50
  packageType?: string; // 50
  volume?: string;
  weight?: string;
  externalPackageCode?: string;
  bidtDeliveryFolderId?: number;
  weightUnit?: string;
  length?: number;
  lengthUnit?: string;
  width?: number;
  widthUnit?: string;
  height?: number;
  heightUnit?: string;
  lengthDefault?: number;
  widthDefault?: number;
  heightDefault?: number;
  unitDefault?: string;
}